requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
boto3==1.42.80
botocore==1.42.80
click==8.3.1
fastapi==0.135.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
s3transfer==0.16.0
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
